+++ web/status-history/src/test/java/net/twisterrob/travel/statushistory/test/HtmlValidator.java
@@ -1,4 +1,4 @@
-package net.twisterrob.travel.statushistory;
+package net.twisterrob.travel.statushistory.test;
 
 import java.io.StringReader;
 import java.io.StringWriter;
